code/requirements.txt
fsspec
pandas


#ann.py
tensorflow
pillow
scipy